Parsing Rework
Took the things I brought over from MCRegionJS, and refactored them to more strongly work with the Java format specifically. Things are a bit more tidy now in that regard.

It does seem like Java regions are the easiest to parse, which is fortunate! It's actually not too bad to iterate over all of the file to get specific chunks and things like that.

Went back to the wiki's docs about the file format to get familiar with it's specifics again.

https://minecraft.fandom.com/wiki/Region_file_format

Part of the rework was that I mainly noticed that since the body of the chunk data is just plain NBT, I can pass that directly into NBTify itself, and have it deal with the compression formats! I wish LCE was structured this way, because that seems to work really well in conjunction with building on top of each section of the file format. By that, I mean that you can read some parts at a time specifically, without having to do the other steps first. You can get the file size and compression format before you have to decompress the actual content, things like that. In LCE chunks, some parts of the compression stages depend on you to have decompressed 'x' stages before you are able to read the header at that point. So, each part of the header is saved in a different section of the compression stage, meaning you have to do each part one by one, rather than being able to query that format information before having to decode the whole content of the chunk.

Have started trying listening to ASMR while coding, and wow it is a great experience. Especially with headphones. I've listened to it before, sometimes to help with sleep. For coding though, it feels like it really helps keep my attention on the project at hand. It is making me sleepy though, but that's a good thing, since I usually tend to stay up coding all night on the weekends, haha.