Chunk Finding
Trying to locate the chunk I used as a demo for NBTify, since I want to add support for reading Anvil/Region files into NBTify afterall. I think it's not too out of scope, because NBTify is already set up to handle other headers like Bedrock Level files, and of course all the compression ones too. If something is plain NBT data, and it just uses a header to describe it, I think I should probably just support that out of the box? I thought it would make sense for it to be separate from NBTify, but maybe it can just be a different module or something. That's actually not a bad idea!

That just opened up a new set of ideas for me! Maybe I can represent different file types using extension classes (objects) of `NBTData`, so it doesn't get cluttered with extra unnecessary properties. So maybe I can make a class like `BedrockLevel` or something, which will have the `version` property on the prototype there, rather than on `NBTData`. Seems a bit nicer come to think of it. And maybe the `.mca` alternative could be `Anvil` or something.

After looking for the specific region I was testing on before, I think it's from a different world save, back from on my Chromebook, not this superflat demo here.