Functional Factor
I want to use more functional data structures to work on the chunks. I realized my original setup wasn't going to work with generators, mostly because of error handling inconsistencies/challenges, and that I want to find whether chunks should be loaded all at once (they shouldn't, that kind of defeats the purpose of them), or instead how I can load individual ones at a time, using APIs on top of the data. It's kind of confusing as to how I want to organize opening each of the formats, and where the line of abstraction should go when working with the different data formats.

I think there should probably be both implementations in terms of the angles of abstraction. That's the most difficult way to do it, but it's the most powerful, and the most extendable I think.

By all of that, I mean, do I build a single API that can handle all of the world formats, all in one go, or (my favored, more difficult option) do I build APIs for each save format, then build a streamlined API on top of those, which bridges the gap for them, so you can use them irrespective of what you're saving to? I like that method a lot, because you only have to learn one set of APIs, and the underlying platform-specific (Bedrock, Java, LCE) implementations will do the more-detailed work.

I think I am going to help commit myself to this idea, as right now I keep trying to build a single API that would work for all of the platforms, all at once, and I don't think that's fullproof, nor something maintainable.

So, from this point, I'm going to keep letting myself know that each platform will have it's own implementation for it's respective save format, and the universal compatibility parts can come later.

This is already the mindset I'm using for NBTify, and that has turned out very nice to what I wanted it to look like. It's probably better than I thought actually, because I wasn't sure if that was even possible to have cross-compatibility between big endian and little endian files. But, once you have two implementations for something, and they both resolve to the same parsed format (JS objects in this case), then you can manipulate them however you want, and just let the export format do it's thing. Exceptional! Abstractions :)

I think the other key thing here too, is that each of the formats is already 1:1 compatible with itself, so just making sure that the parsed formats are the same primitives are the last thing to resolving that they will line up.

I wish that Minecraft itself were to have done this, it would be much better for the game I think. By that, I mean things like texture file names, item/block/entity/biome IDs, NBT tag names, all those sorts of things. If the games work different, that can be separate from the fact that one world can work in both versions. But these game primitives aren't even the same, so you can't use the same world data on both platforms. And they also haven't made their own compatibility layer between these game primitives, which is unfortunate. I think the proper solution is indeed in the works though, to bring Java's 'The Flattening' over to Bedrock. That will make world symmetricality something much easier to accomplish, because you'd just be worrying about the 'physical' save format itself (.mca vs LevelDB, etc). Say if there was a bug with the Llama in Bedrock, that's fine in one set of terms, because that's when you play the game. That doesn't dictate whether that Llama is also a valid Java Llama.

Ok, let's do this.

I also want to debate setting aside the browser requirement, and making these things a Node-only project, but I also think that setting that aside will only make things harder down the line for making it runnable in the browser, so I don't think I want to do that. Maybe my demos can currently only work in Node, but this should definitely be implementable for the browser, the main issues to work around are the file system resolutions for world save file references (and how the API should deal with this), and getting some sort of leveldb-mcpe implementation working in the browser (be it, though WASM, or some other method of simply working with the LevelDB format. My philosophy of discovery has been that you have to remember it's just a data format, so you don't need to specifically use a single program or implementation to open it, you just have to have something which follows the same format to implement that functionality for you. That made no sense, but yeah).