Region Setup
Parsing all of the regions in the test world's folder! It does each chunk asynchronously too! Not sure if this can get out of hand, but it seems to work great on my machine (lol). Gonna have to do some experimenting to see if you can consistently gunzip that many files at once in other places too.