Region Methods
Want to enable managing the content of the Region using methods on the instance, kind of like an Array, but not storing the Chunks sequentially, and instead in a matrix-like way.

I haven't worked on this in a few days, so I kind of forgot what I was working on last. I think these changes were why I had to make edits to NBTify.