Packags Upgrde
Can't really upgrade NBTify in this case, I may go back and re-add support for the `NBTData` format options generic, I think it is useful for me afterall, I'm not quite sure yet. Going to do some more testing for it.

I'm going to look into adding reading and writing support for Java and LCE regions and chunks, and I'm thinking about either combining the two into a single Region reader implementation, rather than managing two separate ones. I think I'm going to do that, since I kind of want to start fresh on both anyways. It will make it easier to make a universal format defined across a single package too, so maybe I will do that next :)

So I'm gonna have an IO package for reading and writing world saves, and that's gonna depend on the separately-defined Region-Types repo, which has all of the documentation for the NBT and chunk properties formats.

Listening to The Harmony Codex for the first time! It's out today!!!
https://www.youtube.com/playlist?list=OLAK5uy_ng3Ruh6KTb_yeqdsM2rOX-RQTj4pNwLyo