Region / Chunk Refactorings
This is also brought over from MCRegionJS.

The coupling between Region and Chunk data was too highly tied together, one depended on the other, and abstracted the other into the other. Now, Regions are more viewed as just containers over the individual Chunk data entries, and reading the Chunk data is now specific to each Chunk and it's own data individually. This is already a big help in designing the mental map for the intermediate format, since each platform does use Chunks, not necessarily Regions. So using Chunks as the bare primitive makes more sense there, because they can be compared evenly across the board, just using objects and types.

Flamingosis :)