Chunk Header Scoping + Re-Write
Going to build a new set of write APIs from the ground up, since the reading APIs have essentially completely changed.

Things are continuing to look very similar to MCRegionJS now, thankfully :)

*Long Distance Call* (woah, just realized that Yuval Ron's creative tone likely stems from 80s Rush! Hadn't made the sonic connection before, just knew it sounded familiar and awesome. And, of course that it's space-themed)

A discussion I started about where the Chunk header should be handled, as part of the Chunk parser, or the Region instead?

"Do you think the Chunk header part in Java's Region files is part of the Chunk's handling, or rather part of the Region's format? Looking into it more, I think it actually may make more sense to do that handling as part of the Region format, as previously I was doing it with the Chunk reading section
https://minecraft.wiki/w/Chunk_format
https://wiki.vg/Region_Files#Chunk_Header
https://minecraft.wiki/w/Java_Edition_Alpha_level_format#Chunk_format
Considering that the header only appears to be there when nested inside of Region files (having thought about it), it seems more streamlined for that to be handled by the Region parsing itself, rather than Chunk parsing
I was curious what Alpha did because it didn't use Regions back then, so I was curious if that header was still around back then too, but turns out it wasn't!
It makes sense where the legacy gzip compression comes from too, as Alpha files used gzip, but modern versions just use deflate (zlib)"