Region Refactorings
Scambot 2, The Madness of Many, Parrhesia, Demiurgs of Plancklengthville, Dog, and Inkling! (in reverse order, that was recently listened to first)

So the goal with this time around is to move the Region-specific Chunk headers into the Region reading results, rather than in the Chunk parsing stage. I also want to dissolve the Locations and Timestamps functions into the single `readEntries()` function, because I want to use a single `Entry` type instead now, one which holds the values for all of the dedicated Region Chunk entry information, rather than only the Chunk data. It more simply couples things together where you only have to do a single read call, and you get all of the metadata for that Chunk entry at once, rather than needing to read each metadata section manually yourself. So I think it's a nice two-way abstraction on top of the Region format, which would allow you to do everything you already could before, but with less manual calls. This is like how I wanted NBTify to work, where you just pass in the specific metadata flags, and the parser/writer will resolve those to the format as necessary, you don't have to handle any of the actual format yourself! This is also the stack I want to follow with world data once these primitives for save data can be established. It's that the metadata is there and nicely laid out for you to manipulate, it's the literal outputs and inputs that you don't need to worry about quite as heavily.

I updated the dependencies to the latest versions too, including NBTify! I'm going to see if needing the `NBTData` generic for format options will come up as necessary or not, as I feel like I need it, but I want to see if it's actually something required or not.

https://minecraft.wiki/w/Region_file_format#Payload

I absolutely love 'Freezer Burn', the ending section is the best.