Entry Value Parsing
Flattening the Chunk parsing into a single call, rather than needing to parse each of the Region metadata entries individually.

I need to re-add in the calculations provided by the Region's Chunks' headers though, as right now it's silently correctly parsing the trailing empty Chunk data bytes correctly, but that's the case because it allows the compressed content to have trailing data, which I don't want to rely on, because that won't work for uncompressed NBT entries, since I don't want to silently allow that in NBTify.

https://minecraft.wiki/w/Region_file_format#Payload

So essentially the error is that this version is only working because all of the Chunks have their content compressed, and that is what's allowing their trailing empty data to not cause errors during the parsing stage.