Nullish Unloaded Chunks
A simpler way of handling unloaded Chunks inside of a Region file. This is how I had things originally, when set up with plain Uint8Array values as the return Entry type. Rather than bringing that nullish handling into the object itself, and having placeholder-ish unset values for the rest of the entry, it just makes more sense for the Entry itself to be null, because those values can't hold anything else when that Chunk is empty anyways.