Tidier Control Flow
This was plaguing me for some time during the debugging of trying to set this up. I wanted to essentially get the same thing you get with guard clauses, but inside of the generator function. This accomplishes that because now it properly exits the loop after also yielding the value of that iteration. That was what I hadn't completely grasped yet, that yielding essentially pauses the function state at that moment, and it's picked up right back at that spot again when called in the next iteration. And it also does that in my other scenarios because I've pretty much always tended to put my yield clauses at the end of the loop body, hence it finalizes it's completion on the next loop, and there just happens to be no content in the loop body after that point.

There should be a neat JS keyword for this scenario! I'm not sure what it would be called, or if it already exists, but it would be `yield-continue` essentially. Maybe this was will do just fine :)

https://stackoverflow.com/questions/45240857/does-javascript-execute-code-after-a-yield-statement

The changes to the console logs are just to demo things more easily in the console, to see things without having to scroll, but still get all of the error messages I might get when I build something wrong.