Flattened Control Flow!
This puts the Chunk Entry reading into a single procedural function call! Yay! And there aren't any weird number magic cases going on anymore.

Looks like the HEADER_LENGTH check might not be necessary afterall? Trying it over all of the Chunk data, it didn't change the success of the parsing to remove the check, so I think it was a check I didn't need to add because at some point along the way I wasn't totally sure where the Chunk data length checks weren't working, so that was there to help fix that, as I thought that's what the solution had to be.