Chunk Class
This is to ensure that all of the properties read during the Region entry parsing stage are propagated down to the write process too. I realized that the original timestamp for the Chunk wasn't brought down to the NBT object representation of the Chunk. As of now, I only want to update the timestamp explicitly, I'm not sure whether my editing tools should update them automatically. I could see it causing issues if a timestamp has to be a certain date range for some reason, say if you wanted to ensure that the Chunks weren't opened by a certain date, this would allow you to only update the timestamp if you explicitly want to.

In reality, I probably will make it so it will optionally leave the original timestamp as-is, so that would be the opt-in behavior instead. I think most users probably would want the dates of the Chunk being read to be updated with every edit. I just want to ensure that you don't have to update it if you specifically don't want to.