Chunk Writing
Not sure if there should be an equivalent of `function* readEntries()` for writing to a Region file. It should probably only be `function writeRegion()` right? I can't think of a scenario where you could conditionally write to a Region using a Generator or something of the sort. I don't think it's really possible/logical to write only a few Chunks to a Region, then come back later to write more. That seems more like a reading stage improvement/feature, which doesn't really apply here.