Chunk Header Compression Byte
Next demo of reading content. Trying to get all of the values in one map iteration, rather than looping over things multiple times. Going to take multiple versions to find what works the best, just like the reading code needed to do.

https://wiki.vg/Region_Files#Chunk_Header

Yeah, turns out the read Chunk bytes were one byte too long when reading from the source Region file, and it's because of this small overlook I missed when implementing the Region logic. The length of the Chunk data for Region Entries includes the compression flag byte, but I remove that when doing the Entry building, so I added the handling for that to make it the one byte shorter as it needs to be. Thankfully the whole Region continued to parse fully correct with no errors, so it was a silent little bug that wasn't causing errors, it just wasn't implemented right. That also works out great because all of the Chunks are compressed, and the trailing bytes always happened to be zeros, meaning zlib just ignored them when decompressing. Had it been plain NBT, it would've failed, and had it been non-zeros I think zlib would have failed too. I might be mis-remembering whether that's the case for zlib, I thought I heard that somewhere though.

Flamingosis; thinking of listening to parts of The Incident after this, I think because I was thinking about my old PE worlds from my phone, and that's when I was listening to that album a lot, when I made those worlds just before starting high school. So essentially I was thinking about Minecraft during that time period, which reminded me about that music.