Locations Calculations
This part is hard! I'm trying to figure out how to map the values for these correctly. I'm also not sure whether the compressed results are exactly the same, and as to whether that could be throwing off the offset values. I think the offsets would be much closer though.