Single-Buffer Region Write
(Concept, not ready for this part yet)
(Why are they so many friken ants on my desk aaaagh)
(I need to build a Bandcamp alternative just in case it goes under, even if it's just for my projects; I'm getting a bit worried about whether it will continue to have longevity into the future with all of this corporate drama and purchases. It will probably be fine, I hope that's the case, that would be the best scenario. Bandcamp is great!)