Offset Accumulation
This doesn't appear to be right, I'm not completely sure about it yet. It's possible some of the Chunks have smaller sizes when being re-saved, but I'm a bit skeptical about that. This is kind of what I ran into yesterday too. So, I guess I am doing a similar set of calculations here still, just with a bit different of organization.