Writing ByteLengths
Realized that my use of ` * LOCATIONS_LENGTH` during the reading stage is actually incorrect. That should be a different constant, it just happens to be the same value of `4096`. The byte length of the Chunk entry is a multiple of `4096`, but not because that's the length of the Locations header, they just happen to be the same number. I need to make a constant called `ENTRY_LENGTH` or something.