Region Entry Edit Sorting
Started understanding from Evergreen that Chunks are ordered in the Entries part of the Region file based on when they were written to the file, with the most recent one being at the end. The 'Locations' and 'Timestamps' part of the file is always statically-ordered, based on the coordinates of the Chunk within the Region itself. I assumed that the Chunk entry data was written to the file in this same pattern, however they are appended/moved to the end of the file when they are saved next, hence why the offsets for the Chunks aren't numerically ordered symmetrically to their coordinates in the Region, it's just based on when they were last written. You could technically recreate this order with the timestamp values, but I think these can overlap across Chunks entries, meaning they could have the same timestamp values, making them inaccurate to sort the entries by.

The main reason for wanting to maintain the original edit order is to be able to diff the byte output of the re-saved Region file to that of the original. With how I was going to save the files previously, they would just be written to the Region with the correct (original) timestamps in the file's header, but their order in the file's entries themselves would go all over the place, meaning the diff wouldn't be something useful to detect changes with.

https://stackoverflow.com/questions/5494974/convert-1d-array-index-to-2d-array-index

```js
const columnCount = 32;

for (let i = 0; i < columnCount * columnCount; i++){
  const row = Math.floor(i / columnCount);
  const column = i % columnCount;
  console.log(`${i}: ${row},${column}`);
}
```

Now I just have to figure out how to get the Region-local coordinates for the Chunks based on it's index in the Locations header. Then I can store that on the Chunk/Entry result itself, and use that to maintain the Chunk's state no matter how the data is sorted.

The reason I have to do that is because for unloaded `null` Chunks, you lose the Chunk's state association with that coordinate, if you sort the array using another method. Currently the Chunk's coordinates are managed by where it's stored in the array, and if you change the order of the array, you lose the association as to which coordinate that unloade value is at. So it still maintains that it's an unloaded Chunk, but you no longer know where that one is at. It's just an arbitrary unloaded item. With the other Chunks, it's ok because you can just parse the coordinates from the NBT if you really needed to, but that's not ideal and that's what I wanted to stray away from in the Region handling anyways.

So this change would make it so the array index specifies where in the Entries order the Chunk data should be saved, while the `x,z` properties on `Chunk` and `Entry` would take over the need for maintaining what spot it should be written to the Locations and Timestamps section (and of course it's location as coordiantes in the Region)
This is a hectic thing.

Now I want to try and write a script to randomize where Chunks are saved within a Region, sounds really neat hehe.

Now I'm also thinking I could still keep the array index for Region-relative coordinates, and instead manage the edit order with a property, that actually sounds much more manageable. Going to look into that next, after committing these other demos and such.

Snarky Puppy - The World is Getting Smaller