Timestamp Discoveries
Looked into it, and yes, Chunks can have multiple of the same timestamp, so it's not something you can rely on to safely sort the edit time indexing of the Entries in the Region.

Listened to some more of The World is Getting Smaller. Now I'm starting Scambot 2. Now I'm 'In The Trees'.