Sorting with Buffers!
Wow! Realized I can still do this sorting setup while still implementing Uint8Array objects with the wrappers for each of the Entries' data. Very helpful! This is because the buffers for those just use the underlying Region file's data, for the read step. And since those are just properties which represent the starts and ends of the view on the bytes, you can just use those as the checks to validate against! I kind of forgot that the value of those is persisted to the final instance of the Uint8Array, it's not just present at the constructor. Kind of obvious looking at it now, but I think I didn't think of it because it's more of an implementation detail as to how I have the Entry views on the Region data, since I wanted to make sure I didn't duplicate any data unnecessarily without reason.