Region Reset
I think I might want to try some fresh things with how I manage the state of the Region file, in regards to reading it's entries and it's read/write order. I also want to see if I can generalize it so it's not Java-specific, since it shares the same shape for both non-Anvil files (plain `.mcr`), and also LCE's Region format, which is also just `.mcr`. Essentially they are all the same thing anyways, since it's the chunk format that changes between all of them, but I want to remove the Chunk implementation part from the Region code itself, it should accept any of those types; probably.