Region Debugging
Looking into trying to figure out how entries are stored/ordered in the Region file, since it doesn't use the coordinates themselves, at least in terms of where the Chunk for those coordinates is saved within the Region file. I think it uses a stack-based thing, where the most recently edited Chunk is tacked on to the end, and the missing gap is simply just copied over cleanly, the rest of the Region doesn't have to be rearead or serialized again.

https://minecraft.wiki/w/Region_file_format

Wow, these came in handly very well, for some reason this week has been a bit tougher lately, it's crazy how many of these are relatable.

https://www.youtube.com/watch?v=dlpezhtrATQ
https://www.youtube.com/watch?v=yDdIijyAn6k
https://www.youtube.com/watch?v=Tl20Ke2Y58g

Dark Empath is something I've never heard of before, nor thought of being possible! Very intriguing.

I want to make this Region parser return a `JavaEntry | LegacyConsoleEntry`, which means the entry header parsing is left for after the chunking step. I think this is a more streamlined approach, and it's better at following the patterns like that which I have been using in my own projects.

If any of this commit message made no sense, it's because I'm almost falling asleep while writing this, getting real tired at night for once, yay!!!

I still have more to do, but this is a good pausing point for the night, and it's a good middleground to commit/track before I make the new official changes to it.