Entry Header Parsing
I want to find a way to merge these entry handling mechanics into the Region file format itself maybe, so you don't have to manage it on multiple levels.

Say like if `readRegion()` returned a `Region<JavaEntry | LegacyConsoleEntry>`, and you would handle each accordingly.

So while the Region format is a format on it's own, it's only ever used with either of those platforms, and it can automatically detect which one it is when you are opening it. It's just based on the header of the entry itself, so it shouldn't be too hard to work with the Region entry blocks and such. The hard part I was running into is recalculating the lengths of things when you rebuild the entry blocks on your own, since they have byte padding and things like that.