Entry Header Parser Merging
Trying to re-embed the header parser logic into the Region implementation, now I just have to figure out how to re-format the block so it can have the correct length and header info when re-writing it. Having to go back to (dadd700) so I can see how I was calculating these header parts.