Symmetrical Test Checking
I think this demo was meant to try running the read/write comparisons for Region entries on their own, outside of the overarching Region handling itself. This was from last night, to which I fell asleep before finishing or committing the progress I had so far for it. I was going to try doing a recap of what I was doing and where I was, but I don't have much time for coding tonight, so I'm going to leave that for another day, possibly tomorrow instead!