Chunk NBT Format Extraction
Realized how I can use persistent structures for reading existing chunks, but also use the validation format object to ensure the incoming data is in the right format, and the outgoing data is correct as well.

Dee 'n' A is one of my favorite songs, come to think of it now, wow I love this one. I think I only just realized right now, just how great it is. And parts of it remind me of Wine and Pickles too I think.