Chunk Subclass Removal
Realized recently that I can also use a custom interface type on top of `NBTData`, rather than extending it as a class. That actually seems like a nice way to go, since it doesn't mean you have to use any constructors or `instanceof` checks. It simply will only allow you to get and set properties on that data object, if it's the type you intend it to be. So say like for how I am using `Object.assign()` in the code here.

My changes in `region.ts` broke how things are handled here, and it's still a bit weird. I have to make the region parsing generic to work with any platform or region version, as I don't want to maintain separate ones for each edition.