Region-Relative Chunk Coords Validation
Looking into whether the indexes are being calculated correctly in reference to the NBT values in the chunks themselves.
So like where the chunk is in the `Chunk[]`, that index number should derive to the same coordinates in the NBT itself, at least for `r.0.0.mca`, where those NBT coordinates happen to be the same as the region-relative coordinates.
Not sure if that makes sense but yeah hehe.