Sectors Distinctions
What should the concept of the Region entries be called? That's the name I'm going with as of now, `Entry`
It seems a little too generic to me though
`Chunk` would be the NBT result itself, hence that `Region` would be `Chunk[]`
So '`Entry`' would be the part of reading the raw file itself
What should that be named?

Another way to put it for the naming
A `Chunk` can be made arbitrarily, it's just an object of the NBT data
A `Region` can be made arbitrarily, it's just a `Chunk[]`
An `Entry` needs to have an associated `Region` to get it's values from, whether it's an existing file, or the other neighboring `Chunk`s to be serialized with

So like instead of reading the location and timestamp entries on our own with the API, it just looks like this
```ts
const data = await readFile("./my-region.mca");
const sectors = readSectors(data);
const region = await readRegion(data, sectors);
```

We wouldn't have to do all of that on our own, I think that's how it should work under the hood though
So like maybe `readRegion()` could just call `readSectors()` itself